Escape file name for html_url & url
Sense Files API is updated to accept encoded file path
as part of the URL for API calls in order to support
special characters. The url & html_url properties of
each fileEntry is populated by vfs-http-adatper and
are heavily referenced by Sense web. Update the logic
for populating these properties to include encoded
file name in the path.