Allow custom response handler
This PR was motivated by a security need to have all files returned JSON
serialized as opposed to raw which exposed an XSS vulnerability. There needed
to be a way to return the full file payload to the response stream in a
way customized by the client. In our specific use case we willprovide a function
that wraps the buffer in json before piping out to the response, but there are other
possiblie scenarios that could occur down the line, hence making this very general
as opposed to a default json wrapper.

See also: PR in CDSW